requirements.txt
pandas
numpy
requests
geopy
timezonefinder
pytz
skyfield
sentence-transformers
transformers
faiss-cpu
scikit-learn
matplotlib
opencv-python
streamlit
streamlit-drawable-canvas
requests-cache
retry-requests
pyngrok
openmeteo_requests
